+++ igreja_dll/WindowsFormsApp1/Formulario/formularioMovimentacaoEntrada/FrmCantina.cs
@@ -35,8 +35,8 @@ namespace WindowsFormsApp1.formulario.formularioMovimentacaoEntrada
             Cantina a = (Cantina)modelo;
             try
             {
-                a.Valor = double.Parse(txtValor.Text);
-                a.Valor = double.Parse(a.Valor.ToString("F2"));
+                a.Valor = decimal.Parse(txtValor.Text);
+                a.Valor = decimal.Parse(a.Valor.ToString("F2"));
             }
             catch { MessageBox.Show("digite numeros"); txtValor.Text = ""; }
         }
