+++ igreja_dll/WindowsFormsApp1/Formulario/formularioMovimentacaoEntrada/FrmDizimo.cs
@@ -32,8 +32,8 @@ namespace WindowsFormsApp1.formulario.formularioMovimentacaoEntrada
             Dizimo a = (Dizimo)modelo;
             try
             {
-                a.Valor = double.Parse(txtValor.Text);
-                a.Valor = double.Parse(a.Valor.ToString("F2"));
+                a.Valor = decimal.Parse(txtValor.Text);
+                a.Valor = decimal.Parse(a.Valor.ToString("F2"));
             }
             catch { MessageBox.Show("digite numeros"); txtValor.Text = ""; }
         }
